Add Google map

+++ app/src/main/java/com/islam/cities/data/model/Coord.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Coord(
     @SerializedName("lon")
-    val lon: Double? = null,
+    val lon: Float? = null,
     @SerializedName("lat")
-    val lat: Double? = null
+    val lat: Float? = null
 )
